refator: duplicate email input component Link T-12340
